Use Py_ssize_t for PySet_Size() like all the other Py*_Size() functions.

+++ Include/setobject.h
@@ -76,7 +76,7 @@ PyAPI_DATA(PyTypeObject) PyFrozenSet_Type;
 
 PyAPI_FUNC(PyObject *) PySet_New(PyObject *);
 PyAPI_FUNC(PyObject *) PyFrozenSet_New(PyObject *);
-PyAPI_FUNC(int) PySet_Size(PyObject *anyset);
+PyAPI_FUNC(Py_ssize_t) PySet_Size(PyObject *anyset);
 #define PySet_GET_SIZE(so) (((PySetObject *)(so))->used)
 PyAPI_FUNC(int) PySet_Contains(PyObject *anyset, PyObject *key);
 PyAPI_FUNC(int) PySet_Discard(PyObject *set, PyObject *key);
